+++ src/main/java/com/lmax/disruptor/AggregateEventHandler.java
@@ -21,7 +21,7 @@ package com.lmax.disruptor;
  * @param <T> event implementation storing the data for sharing during exchange or parallel coordination of an event.
  */
 public final class AggregateEventHandler<T>
-    implements EventHandler<T>, LifecycleAware
+    implements EventHandler<T>
 {
     private final EventHandler<T>[] eventHandlers;