+++ src/main/java/com/lmax/disruptor/BatchEventProcessor.java
@@ -24,9 +24,6 @@ import static com.lmax.disruptor.RewindAction.REWIND;
  * Convenience class for handling the batching semantics of consuming entries from a {@link RingBuffer}
  * and delegating the available events to an {@link EventHandler}.
  *
- * <p>If the {@link EventHandler} also implements {@link LifecycleAware} it will be notified just after the thread
- * is started and just before the thread is shutdown.
- *
  * @param <T> event implementation storing the data for sharing during exchange or parallel coordination of an event.
  */
 public final class BatchEventProcessor<T>
@@ -209,7 +206,6 @@ public final class BatchEventProcessor<T>
                     break;
                 }
             }
-
             catch (final Throwable ex)
             {
                 handleEventException(ex, nextSequence, event);