+++ src/main/java/com/lmax/disruptor/EventHandler.java
@@ -21,7 +21,7 @@ package com.lmax.disruptor;
  * @param <T> event implementation storing the data for sharing during exchange or parallel coordination of an event.
  * @see BatchEventProcessor#setExceptionHandler(ExceptionHandler) if you want to handle exceptions propagated out of the handler.
  */
-public interface EventHandler<T> extends BatchStartAware, LifecycleAware, SequenceReportingEventHandler<T>, TimeoutHandler
+public interface EventHandler<T>
 {
     /**
      * Called when a publisher has published an event to the {@link RingBuffer}.  The {@link BatchEventProcessor} will
@@ -37,4 +37,54 @@ public interface EventHandler<T> extends BatchStartAware, LifecycleAware, Sequen
      * @throws Exception if the EventHandler would like the exception handled further up the chain.
      */
     void onEvent(T event, long sequence, boolean endOfBatch) throws Exception;
+
+    /**
+     * Invoked by {@link BatchEventProcessor} prior to processing a batch of events
+     *
+     * @param batchSize the size of the batch that is starting
+     */
+    default void onBatchStart(long batchSize)
+    {
+    }
+
+    /**
+     * Called once on thread start before first event is available.
+     */
+    default void onStart()
+    {
+    }
+
+    /**
+     * Called once just before the event processing thread is shutdown.
+     *
+     * <p>Sequence event processing will already have stopped before this method is called. No events will
+     * be processed after this message.
+     */
+    default void onShutdown()
+    {
+    }
+
+    /**
+     *  Used by the {@link BatchEventProcessor} to set a callback allowing the {@link EventHandler} to notify
+     *  when it has finished consuming an event if this happens after the {@link EventHandler#onEvent(Object, long, boolean)} call.
+     *
+     *  <p>Typically this would be used when the handler is performing some sort of batching operation such as writing to an IO
+     *  device; after the operation has completed, the implementation should call {@link Sequence#set} to update the
+     *  sequence and allow other processes that are dependent on this handler to progress.
+     *
+     * @param sequenceCallback callback on which to notify the {@link BatchEventProcessor} that the sequence has progressed.
+     */
+    default void setSequenceCallback(Sequence sequenceCallback)
+    {
+    }
+
+    /**
+     * Invoked when a {@link BatchEventProcessor}'s {@link WaitStrategy} throws a {@link TimeoutException}.
+     *
+     * @param sequence - the last processed sequence.
+     * @throws Exception if the implementation is unable to handle this timeout.
+     */
+    default void onTimeout(long sequence) throws Exception
+    {
+    }
 }