+++ src/perftest/java/com/lmax/disruptor/support/ValueMutationEventHandler.java
@@ -15,13 +15,12 @@
  */
 package com.lmax.disruptor.support;
 
-import com.lmax.disruptor.BatchStartAware;
 import com.lmax.disruptor.EventHandler;
 import com.lmax.disruptor.util.PaddedLong;
 
 import java.util.concurrent.CountDownLatch;
 
-public final class ValueMutationEventHandler implements EventHandler<ValueEvent>, BatchStartAware
+public final class ValueMutationEventHandler implements EventHandler<ValueEvent>
 {
     private final Operation operation;
     private final PaddedLong value = new PaddedLong();