+++ CHANGELOG.md
@@ -6,6 +6,10 @@
     - Minimum Java version now 11
     - Issue #323 - `WorkerPool` and `WorkProcessor` have been removed, no more `Disruptor::handleEventsWithWorkerPool`
     - `Disruptor` constructors using Executor have been removed. Use ThreadFactory instead.
+    - Rolled up event handling extnsion interfaces on to EventHandler:
+      - `BatchStartAware`
+      - `LifecycleAware`
+      - `SequenceReportingEventHandler`
  - Deprecated ThreadHints.onSpinWait()
  - FatalExceptionHandler and IgnoreExceptionHandler now use the JDK 9 Platform Logging API, i.e. System.Logger
  - Add rewind batch feature to the BatchEventProcessor